Fixed broken tests; simpler bootstrap setup
