+++ test/TestBuildingBlocks/BaseIntegrationTestContext.cs
@@ -1,16 +1,12 @@
 using System;
-using System.Collections.Generic;
 using System.Net.Http;
-using System.Net.Http.Headers;
 using System.Threading.Tasks;
-using JsonApiDotNetCore.Middleware;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
 
 namespace TestBuildingBlocks
 {
@@ -23,7 +19,7 @@ namespace TestBuildingBlocks
     /// <typeparam name="TStartup">The server Startup class, which can be defined in the test project.</typeparam>
     /// <typeparam name="TRemoteStartup">The base class for <typeparamref name="TStartup"/>, which MUST be defined in the API project.</typeparam>
     /// <typeparam name="TDbContext">The EF Core database context, which can be defined in the test project.</typeparam>
-    public abstract class BaseIntegrationTestContext<TStartup, TRemoteStartup, TDbContext> : IDisposable
+    public abstract class BaseIntegrationTestContext<TStartup, TRemoteStartup, TDbContext> : IntegrationTest, IDisposable
         where TStartup : class
         where TRemoteStartup : class
         where TDbContext : DbContext
@@ -34,6 +30,11 @@ namespace TestBuildingBlocks
         private Action<IServiceCollection> _afterServicesConfiguration;
 
         public WebApplicationFactory<TRemoteStartup> Factory => _lazyFactory.Value;
+
+        protected override HttpClient CreateClient()
+        {
+            return Factory.CreateClient();
+        }
 
         protected BaseIntegrationTestContext()
         {
@@ -101,102 +102,6 @@ namespace TestBuildingBlocks
             var dbContext = scope.ServiceProvider.GetRequiredService<TDbContext>();
 
             await asyncAction(dbContext);
-        }
-
-        public async Task<(HttpResponseMessage httpResponse, TResponseDocument responseDocument)>
-            ExecuteGetAsync<TResponseDocument>(string requestUrl,
-                IEnumerable<MediaTypeWithQualityHeaderValue> acceptHeaders = null)
-        {
-            return await ExecuteRequestAsync<TResponseDocument>(HttpMethod.Get, requestUrl, null, null, acceptHeaders);
-        }
-
-        public async Task<(HttpResponseMessage httpResponse, TResponseDocument responseDocument)>
-            ExecutePostAsync<TResponseDocument>(string requestUrl, object requestBody,
-                string contentType = HeaderConstants.MediaType,
-                IEnumerable<MediaTypeWithQualityHeaderValue> acceptHeaders = null)
-        {
-            return await ExecuteRequestAsync<TResponseDocument>(HttpMethod.Post, requestUrl, requestBody, contentType,
-                acceptHeaders);
-        }
-
-        public async Task<(HttpResponseMessage httpResponse, TResponseDocument responseDocument)>
-            ExecutePatchAsync<TResponseDocument>(string requestUrl, object requestBody,
-                string contentType = HeaderConstants.MediaType,
-                IEnumerable<MediaTypeWithQualityHeaderValue> acceptHeaders = null)
-        {
-            return await ExecuteRequestAsync<TResponseDocument>(HttpMethod.Patch, requestUrl, requestBody, contentType,
-                acceptHeaders);
-        }
-
-        public async Task<(HttpResponseMessage httpResponse, TResponseDocument responseDocument)>
-            ExecuteDeleteAsync<TResponseDocument>(string requestUrl, object requestBody = null,
-                string contentType = HeaderConstants.MediaType,
-                IEnumerable<MediaTypeWithQualityHeaderValue> acceptHeaders = null)
-        {
-            return await ExecuteRequestAsync<TResponseDocument>(HttpMethod.Delete, requestUrl, requestBody, contentType,
-                acceptHeaders);
-        }
-
-        private async Task<(HttpResponseMessage httpResponse, TResponseDocument responseDocument)>
-            ExecuteRequestAsync<TResponseDocument>(HttpMethod method, string requestUrl, object requestBody,
-                string contentType, IEnumerable<MediaTypeWithQualityHeaderValue> acceptHeaders)
-        {
-            var request = new HttpRequestMessage(method, requestUrl);
-            string requestText = SerializeRequest(requestBody);
-
-            if (!string.IsNullOrEmpty(requestText))
-            {
-                request.Content = new StringContent(requestText);
-
-                if (contentType != null)
-                {
-                    request.Content.Headers.ContentType = MediaTypeHeaderValue.Parse(contentType);
-                }
-            }
-
-            using HttpClient client = Factory.CreateClient();
-
-            if (acceptHeaders != null)
-            {
-                foreach (var acceptHeader in acceptHeaders)
-                {
-                    client.DefaultRequestHeaders.Accept.Add(acceptHeader);
-                }
-            }
-
-            HttpResponseMessage responseMessage = await client.SendAsync(request);
-
-            string responseText = await responseMessage.Content.ReadAsStringAsync();
-            var responseDocument = DeserializeResponse<TResponseDocument>(responseText);
-
-            return (responseMessage, responseDocument);
-        }
-
-        private string SerializeRequest(object requestBody)
-        {
-            return requestBody == null
-                ? null
-                : requestBody is string stringRequestBody
-                    ? stringRequestBody
-                    : JsonConvert.SerializeObject(requestBody);
-        }
-
-        private TResponseDocument DeserializeResponse<TResponseDocument>(string responseText)
-        {
-            if (typeof(TResponseDocument) == typeof(string))
-            {
-                return (TResponseDocument)(object)responseText;
-            }
-
-            try
-            {
-                return JsonConvert.DeserializeObject<TResponseDocument>(responseText,
-                    IntegrationTestConfiguration.DeserializationSettings);
-            }
-            catch (JsonException exception)
-            {
-                throw new FormatException($"Failed to deserialize response body to JSON:\n{responseText}", exception);
-            }
         }
 
         private sealed class IntegrationTestWebApplicationFactory : WebApplicationFactory<TRemoteStartup>
